services/pharmacy/pyproject.toml
# Pharmacy — bounded-context service (Phase 5).
# Specification Pattern + Anti-Corruption Layer to RxNav (ADR 0006, 0007).

[project]
name = "pharmacy"
version = "0.5.0"
description = "SmartClinic — Pharmacy bounded context (Specification Pattern + RxNav ACL)."
requires-python = ">=3.12"
dependencies = [
    "shared_kernel",
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.30",
    "sqlalchemy>=2.0,<3.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pydantic-settings>=2.4",
    "aio-pika>=9.4",
    "httpx>=0.27",
]

[tool.uv.sources]
shared_kernel = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pharmacy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
